+++ pkg/apis/scorecard/v1alpha1/types.go
@@ -105,6 +105,16 @@ type ScorecardOutputList struct {
 	Items           []ScorecardOutput `json:"items"`
 }
 
+func NewScorecardOutput() *ScorecardOutput {
+	return &ScorecardOutput{
+		// The TypeMeta is mandatory because it is used to distinguish the versions (v1alpha1 and v1alpha2)
+		TypeMeta: metav1.TypeMeta{
+			Kind:       "ScorecardOutput",
+			APIVersion: "osdk.openshift.io/v1alpha1",
+		},
+	}
+}
+
 func init() {
 	SchemeBuilder.Register(&ScorecardOutput{}, &ScorecardOutputList{})
 }